+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changes
 
 - Drop support for Python 2.6.
 
+- Stop depending on ZODB for anything except testing.
+
 4.1.0 (2014-12-27)
 ------------------